requirements.txt
asgiref==3.8.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
